moved extensions into subfolder
