+++ Extensions/Announcers/GithubAnnouncements/Extensions/BotForkExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Jabbot;
 
-namespace GithubAnnouncements
+namespace GithubAnnouncements.Extensions
 {
     public static class BotForkExtensions
     {
