+++ Extensions/Announcers/GithubAnnouncements/GithubAnnouncements.csproj
@@ -46,16 +46,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BotCommitExtensions.cs" />
-    <Compile Include="BotForkExtensions.cs" />
-    <Compile Include="BotIssuesExtensions.cs" />
-    <Compile Include="BotPullRequestExtensions.cs" />
-    <Compile Include="BotWatcherExtensions.cs" />
+    <Compile Include="Extensions\BotCommitExtensions.cs" />
+    <Compile Include="Extensions\BotForkExtensions.cs" />
+    <Compile Include="Extensions\BotIssuesExtensions.cs" />
+    <Compile Include="Extensions\BotPullRequestExtensions.cs" />
+    <Compile Include="Extensions\BotWatcherExtensions.cs" />
+    <Compile Include="Extensions\SettingsExtensions.cs" />
+    <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="GithubAnnouncer.cs" />
     <Compile Include="GitHub.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SettingsExtensions.cs" />
-    <Compile Include="StringExtensions.cs" />
     <Compile Include="Tasks\CommitTask.cs" />
     <Compile Include="Tasks\ForksTask.cs" />
     <Compile Include="Tasks\IGithubTask.cs" />
